requirements.txt
async-generator==1.10
attrs==22.1.0
black==22.8.0
certifi==2022.9.14
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
distlib==0.3.6
filelock==3.8.0
h11==0.13.0
identify==2.5.5
idna==3.4
iniconfig==1.1.1
mypy-extensions==0.4.3
nodeenv==1.7.0
outcome==1.2.0
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
py==1.11.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.3
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
selenium==4.4.3
sniffio==1.3.0
sortedcontainers==2.4.0
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
virtualenv==20.16.5
webdriver-manager==3.8.3
wsproto==1.2.0
